+++ crl/crl_x509/crl.go
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// Package x509 parses X.509-encoded keys and certificates.
-package x509
+// Package crl_x509 parses X.509-encoded certificate revocation lists.
+package crl_x509
 
 import (
 	"bytes"
