+++ crl/crl_x509/parser_test.go
@@ -1,7 +1,7 @@
 // Copyright 2021 The Go Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
-package x509
+package crl_x509
 
 import (
 	"encoding/asn1"
